+++ py21cmsense/antpos.py
@@ -15,7 +15,10 @@ from . import yaml
 @yaml.yaml_func()
 @un.quantity_input(equivalencies=tp.time_as_distance)
 def hera(
-    hex_num, separation: tp.Length = 14 * un.m, dl: Optional[tp.Length] = None
+    hex_num: int,
+    separation: tp.Length = 14 * un.m,
+    split_core: bool = False,
+    outriggers: bool = False,
 ) -> tp.Meters:
     """
     Produce a simple regular hexagonal array.
@@ -30,29 +33,76 @@ def hera(
         The distance between antennas along a side.
         May have units of distance or time, the latter interpreted as a distance travelled
         by light.
-    dl
-        The distance between rows of antennas.
-        May have units of distance or time, the latter interpreted as a distance travelled
-        by light. If not provided, assume sin(60) * separation (i.e. equilateral triangles).
+    split_core
+        Whether to split the hex into three offset segments.
+    outriggers
+        Whether to add outrigger antennas.
 
     Returns
     -------
     antpos
         A 2D array of antenna positions, shape ``(Nants, 3)``.
     """
-    if dl is None:
-        dl = np.sin(60) * separation
+    sep = separation.to_value("m")
 
-    separation = separation.to_value("m")
-    dl = dl.to_value("m")
+    # construct the main hexagon
+    positions = []
+    for row in range(hex_num - 1, -hex_num + split_core, -1):
+        # adding split_core deletes a row if it's true
+        for col in range(2 * hex_num - abs(row) - 1):
+            x_pos = sep * ((2 - (2 * hex_num - abs(row))) / 2 + col)
+            y_pos = row * sep * np.sqrt(3) / 2
+            positions.append([x_pos, y_pos, 0])
 
-    antpos = []
-    cen_z = 0
-    for row in np.arange(hex_num):
-        for cen_x in np.arange((2 * hex_num - 1) - row):
-            dx = row / 2.0
-            antpos.append(((cen_x + dx) * separation, row * dl, cen_z))
-            if row != 0:
-                antpos.append(((cen_x + dx) * separation, -row * dl, cen_z))
+    # basis vectors (normalized to sep)
+    up_right = sep * np.asarray([0.5, np.sqrt(3) / 2, 0])
+    up_left = sep * np.asarray([-0.5, np.sqrt(3) / 2, 0])
 
-    return np.array(antpos) * un.m
+    # split the core if desired
+    if split_core:
+        new_pos = []
+        for pos in positions:
+            # find out which sector the antenna is in
+            theta = np.arctan2(pos[1], pos[0])
+            if pos[0] == 0 and pos[1] == 0:
+                new_pos.append(pos)
+            elif -np.pi / 3 < theta < np.pi / 3:
+                new_pos.append(np.asarray(pos) + (up_right + up_left) / 3)
+            elif np.pi / 3 <= theta < np.pi:
+                new_pos.append(np.asarray(pos) + up_left - (up_right + up_left) / 3)
+            else:
+                new_pos.append(pos)
+        # update the positions
+        positions = new_pos
+
+    # add outriggers if desired
+    if outriggers:
+        # The specific displacements of the outrigger sectors are
+        # designed specifically for redundant calibratability and
+        # "complete" uv-coverage, but also to avoid specific
+        # obstacles on the HERA site (e.g. a road to a MeerKAT antenna)
+        exterior_hex_num = outriggers + 2
+        for row in range(exterior_hex_num - 1, -exterior_hex_num, -1):
+            for col in range(2 * exterior_hex_num - abs(row) - 1):
+                x_pos = (
+                    ((2 - (2 * exterior_hex_num - abs(row))) / 2 + col)
+                    * sep
+                    * (hex_num - 1)
+                )
+                y_pos = row * sep * (hex_num - 1) * np.sqrt(3) / 2
+                theta = np.arctan2(y_pos, x_pos)
+                if np.sqrt(x_pos**2 + y_pos**2) > sep * (hex_num + 1):
+                    if 0 < theta <= 2 * np.pi / 3 + 0.01:
+                        positions.append(
+                            np.asarray([x_pos, y_pos, 0]) - 4 * (up_right + up_left) / 3
+                        )
+                    elif 0 >= theta > -2 * np.pi / 3:
+                        positions.append(
+                            np.asarray([x_pos, y_pos, 0]) - 2 * (up_right + up_left) / 3
+                        )
+                    else:
+                        positions.append(
+                            np.asarray([x_pos, y_pos, 0]) - 3 * (up_right + up_left) / 3
+                        )
+
+    return np.array(positions) * un.m